requirements.txt
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
distro==1.9.0
exceptiongroup==1.3.1
fs==2.4.16
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
inquirerpy==0.3.4
jiter==0.13.0
markdown-it-py==4.0.0
mdurl==0.1.2
openai==2.28.0
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079158/work
pdfminer.six==20260107
pfzy==0.3.4
pluggy==1.6.0
prompt_toolkit==3.0.52
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.12.1
PyNaCl==1.6.2
PyPDF2==3.0.1
pytest==9.0.3
python-dotenv==1.2.2
requests==2.32.5
rich==14.3.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tomli==2.4.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
